feat: add scan upload service
